+++ api-augment/polkadot/registry.ts
@@ -1,5 +1,5 @@
 // Copyright 2017-2022 @polkadot/api-augment authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import 'https://deno.land/x/polkadot@0.0.4-9/types-augment/registry/substrate.ts';
-import 'https://deno.land/x/polkadot@0.0.4-9/types-augment/registry/polkadot.ts';
+import 'https://deno.land/x/polkadot/types-augment/registry/substrate.ts';
+import 'https://deno.land/x/polkadot/types-augment/registry/polkadot.ts';