+++ api-derive/staking/stakerExposure.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.5';
-import type { EraIndex } from 'https://deno.land/x/polkadot@0.0.4-9/types/interfaces/index.ts';
+import type { EraIndex } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 import type { DeriveApi } from '../types.ts';
 import type { DeriveEraValidatorExposure, DeriveStakerExposure } from './types.ts';