+++ types/interfaces/aura/types.ts
@@ -1,7 +1,7 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { Struct, u64 } from 'https://deno.land/x/polkadot@0.0.4-9/types-codec/mod.ts';
+import type { Struct, u64 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
 
 /** @name RawAuraPreDigest */
 export interface RawAuraPreDigest extends Struct {