+++ types/metadata/util/validateTypes.ts
@@ -1,9 +1,9 @@
 // Copyright 2017-2022 @polkadot/types authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { Registry } from 'https://deno.land/x/polkadot@0.0.4-9/types-codec/types/index.ts';
+import type { Registry } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
 
-import { logger } from 'https://deno.land/x/polkadot@0.0.4-9/util/mod.ts';
+import { logger } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { extractTypes } from './extractTypes.ts';
 import { flattenUniq } from './flattenUniq.ts';