+++ types/types/index.ts
@@ -1,16 +1,17 @@
 // Copyright 2017-2022 @polkadot/types authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+// augment our internal Lookup & Registry interfaces
 import './augmentLookup.ts';
 import './augmentRegistry.ts';
 
 // augmented exports
-export * from 'https://deno.land/x/polkadot@0.0.4-9/types/types/registry.ts';
+export * from 'https://deno.land/x/polkadot/types/types/registry.ts';
 
 // used inside augmented definitions
 export type { Observable } from 'https://esm.sh/rxjs@7.5.5';
 
-// normal exports
+// other exports
 export * from '../create/types.ts';
 export * from './calls.ts';
 export * from './codec.ts';